+++ admin/metabox/ui/components/Forms/ResourcesForm.js
@@ -2,6 +2,7 @@ import React, { Fragment, useContext } from 'react';
 
 import FullWidthToggle from 'metabox/components/Forms/Toggles/FullWidthToggle';
 import TabbedForm from 'metabox/components/Forms/TabbedForm/TabbedForm';
+
 import { AdminContext } from 'metabox/context/adminContext';
 import { handleChange } from 'metabox/utils/event-handlers';