+++ admin/metabox/ui/utils/block-titles.js
@@ -1,0 +1,6 @@
+export const formatBlockType = type => {
+  const trimmed = type.replace( 'gpalab-', '' );
+  const spaced = trimmed.replace( '-', '' );
+
+  return `Untitled ${spaced} block`;
+};